Create EX(14)
Write a C Function that return the addition or subtraction or 
multiplication or division for two numbers. The function should take the  required operation and two numbers as arguments. It also should check that the input operation is one of those operation that mentioned before and if not it should return error. The function should be implemented using switch case.